VideoLingo/backend/requirements.txt
fastapi
uvicorn
yt-dlp
google-genai
edge-tts
python-multipart
pydantic
python-dotenv
